+++ NEWS.md
@@ -3,10 +3,13 @@
 * Update `by_strata_DocTopic`
 * Make examples runnable (thank you [Chung-hong Chan](https://github.com/chainsawriot) for the suggestion!)
 * Speed up (about 15% faster)
+* `save_fig` function
 
 ### Bug fix
 * `weightedLDA` without specifying the number of iterations
 * log-likelihood of dynamic models 
+* saving figures
+* topic labels when there is no keyword topic
 
 # keyATM 0.1.0
 ### Major changes